minor edit on variable 🧻 🔥 ✍  😎 😂

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -10,7 +10,7 @@
 
 size_t print_list(const list_t *h)
 {
-	int node_count = 1;
+	size_t node_count = 1;
 
 	/*return 0 as number of nodes*/
 	if (h == NULL)
